+++ server-mock/server/api-v3/stakewise/test/validators-meta_test.go
@@ -52,9 +52,9 @@ func TestGetValidatorsMeta(t *testing.T) {
 	data := runGetValidatorsMetaRequest(t, session)
 
 	// Make sure the response is correct
-	require.Equal(t, data.Registered, uint(0))
+	require.Equal(t, data.Registered, 0)
 	require.Equal(t, data.Max, vault.MaxValidatorsPerUser)
-	require.Equal(t, data.Available, uint(vault.MaxValidatorsPerUser))
+	require.Equal(t, data.Available, vault.MaxValidatorsPerUser)
 	t.Logf("Received correct response -  active: %d, max: %d, available: %d", data.Registered, data.Max, data.Available)
 }
